# wallet-service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a wallet service.",
        "title": "Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/wallet": {
            "post": {
                "description": "Creates a new deposit or withdraw operation for a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Create a wallet operation (deposit/withdraw)",
                "parameters": [
                    {
                        "description": "Operation Request",
                        "name": "operation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WalletOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.OperationCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/wallets": {
            "post": {
                "description": "Creates a new wallet with an initial balance of 0",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Create a new wallet",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WalletCreateResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/wallets/{walletId}": {
            "get": {
                "description": "Retrieves the current balance of a wallet by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get wallet balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID (UUIDv4)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WalletBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/wallets/{walletId}/operations/{operationId}": {
            "get": {
                "description": "Retrieves the status of a specific operation for a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Get operation status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID (UUIDv4)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operation ID (UUIDv4)",
                        "name": "operationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperationStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.OperationCreateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "operationId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.OperationStatusResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "operationId": {
                    "type": "string"
                },
                "operationType": {
                    "type": "string"
                },
                "processedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.WalletBalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.WalletCreateResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.WalletOperationRequest": {
            "type": "object",
            "required": [
                "amount",
                "operationType",
                "walletId"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "operationType": {
                    "type": "string",
                    "enum": [
                        "DEPOSIT",
                        "WITHDRAW"
                    ]
                },
                "walletId": {
                    "type": "string"
                }
            }
        }
    }
}